fixes lint error

+++ src/scip/benderscut_int.c
@@ -154,7 +154,7 @@ SCIP_RETCODE computeStandardIntegerOptCut(
 
    SCIPdebugMsg(masterprob, "Subproblem %d - Objective Value: Stored - %g Orig Obj - %g\n", probnumber,
       SCIPbendersGetSubproblemObjval(benders, probnumber),
-      SCIPgetSolOrigObj(subproblem, subprobsol)*SCIPgetObjsense(subproblem));
+      SCIPgetSolOrigObj(subproblem, subprobsol)*(int)SCIPgetObjsense(subproblem));
 
    nvars = SCIPgetNVars(masterprob);
    vars = SCIPgetVars(masterprob);
@@ -330,7 +330,7 @@ SCIP_RETCODE generateAndApplyBendersIntegerCuts(
    {
       (*result) = SCIP_FEASIBLE;
       SCIPdebugMsg(masterprob, "No <%s> cut added. Current Master Problem Obj: %g\n", BENDERSCUT_NAME,
-         SCIPgetSolOrigObj(masterprob, NULL)*SCIPgetObjsense(masterprob));
+         SCIPgetSolOrigObj(masterprob, NULL)*(int)SCIPgetObjsense(masterprob));
       return SCIP_OKAY;
    }
